+++ example/lib/screens/components_screen.dart
@@ -1,6 +1,7 @@
 import 'package:example/screens/bottom_navigation_showcase_screen.dart';
 import 'package:example/screens/button_showcase_screen.dart';
 import 'package:example/screens/calculator_chip_showcase_screen.dart';
+import 'package:example/screens/check_row_item_showcase_screen.dart';
 import 'package:example/screens/chip_showcase_screen.dart';
 import 'package:example/screens/display_area_showcase_screen.dart';
 import 'package:example/screens/geometry_area_showcase_screen.dart';
@@ -166,6 +167,12 @@ class ComponentsScreen extends StatelessWidget {
                 context,
                 'Search Row Item Component',
                 const SearchRowItemShowcaseScreen(),
+              ),
+              const SizedBox(height: CoreSpacing.space4),
+              _buildShowcaseButton(
+                context,
+                'Check Row Item Component',
+                const CheckRowItemShowcaseScreen(),
               ),
               // Add more component showcases here as they become available
             ],